+++ sdk/ai/azure-ai-agents/tests/test_agents_client_async.py
@@ -78,7 +78,7 @@ from test_agents_client_base import (
     TestAgentClientBase,
     agentClientPreparer,
     fetch_current_datetime_recordings,
-    fetch_current_datetime_live
+    fetch_current_datetime_live,
 )
 
 # TODO clean this up / get rid of anything not in use
@@ -2150,8 +2150,7 @@ class TestAgentClientAsync(TestAgentClientBase):
                 )
             ]
         vector_store = await ai_client.vector_stores.create_and_poll(
-            file_ids=file_ids, data_sources=ds, name="my_vectorstore",
-            polling_interval=self._sleep_time()
+            file_ids=file_ids, data_sources=ds, name="my_vectorstore", polling_interval=self._sleep_time()
         )
         assert vector_store.id
         await self._test_file_search(ai_client, vector_store, file_id, streaming)
@@ -2198,7 +2197,8 @@ class TestAgentClientAsync(TestAgentClientBase):
                 asset_type=VectorStoreDataSourceAssetType.URI_ASSET,
             )
         vector_store = await ai_client.vector_stores.create_and_poll(
-            file_ids=[], name="sample_vector_store", polling_interval=self._sleep_time())
+            file_ids=[], name="sample_vector_store", polling_interval=self._sleep_time()
+        )
         assert vector_store.id
         vector_store_file = await ai_client.vector_store_files.create(
             vector_store_id=vector_store.id, data_source=ds, file_id=file_id
@@ -2252,11 +2252,11 @@ class TestAgentClientAsync(TestAgentClientBase):
                 )
             ]
         vector_store = await ai_client.vector_stores.create_and_poll(
-            file_ids=[], name="sample_vector_store", polling_interval=self._sleep_time())
+            file_ids=[], name="sample_vector_store", polling_interval=self._sleep_time()
+        )
         assert vector_store.id
         vector_store_file_batch = await ai_client.vector_store_file_batches.create_and_poll(
-            vector_store_id=vector_store.id, data_sources=ds, file_ids=file_ids,
-            polling_interval=self._sleep_time()
+            vector_store_id=vector_store.id, data_sources=ds, file_ids=file_ids, polling_interval=self._sleep_time()
         )
         assert vector_store_file_batch.id
         await self._test_file_search(ai_client, vector_store, file_id, streaming)
@@ -2778,8 +2778,7 @@ class TestAgentClientAsync(TestAgentClientBase):
                 )
             ]
             vector_store = await ai_client.vector_stores.create_and_poll(
-                file_ids=[], data_sources=ds, name="my_vectorstore",
-                polling_interval=self._sleep_time()
+                file_ids=[], data_sources=ds, name="my_vectorstore", polling_interval=self._sleep_time()
             )
             # vector_store = await ai_client.vector_stores.get('vs_M9oxKG7JngORHcYNBGVZ6Iz3')
             assert vector_store.id
@@ -2957,7 +2956,7 @@ class TestAgentClientAsync(TestAgentClientBase):
         """Test using the AzureAISearchTool with an agent."""
         azure_search_tool = self._get_azure_ai_search_tool(**kwargs)
         async with self.create_client(by_endpoint=True, **kwargs) as client:
-            assert isinstance(client, AgentsClient)    
+            assert isinstance(client, AgentsClient)
 
             await self._do_test_tool(
                 client=client,
@@ -2970,7 +2969,7 @@ class TestAgentClientAsync(TestAgentClientBase):
                 uri_annotation=MessageTextUrlCitationDetails(
                     url="www.microsoft.com",
                     title="product_info_7.md",
-                ) 
+                ),
             )
 
     @agentClientPreparer()
@@ -2990,7 +2989,7 @@ class TestAgentClientAsync(TestAgentClientBase):
                 uri_annotation=MessageTextUrlCitationDetails(
                     url="www.microsoft.com",
                     title="product_info_7.md",
-                ) 
+                ),
             )
 
     @agentClientPreparer()
@@ -3093,7 +3092,7 @@ class TestAgentClientAsync(TestAgentClientBase):
                 )
             finally:
                 await client.delete_agent(connected_agent.connected_agent.id)
-    
+
     async def _get_file_search_tool_and_file_id(self, client, **kwargs):
         """Helper method to get the file search tool."""
         ds = [
@@ -3134,8 +3133,8 @@ class TestAgentClientAsync(TestAgentClientBase):
                         text="test",
                         file_citation=MessageTextFileCitationDetails(
                             file_id=file_id,
-                        )
-                    )
+                        ),
+                    ),
                 )
             finally:
                 await client.vector_stores.delete(file_search_tool.resources.file_search.vector_store_ids[0])
@@ -3162,21 +3161,23 @@ class TestAgentClientAsync(TestAgentClientBase):
                         text="test",
                         file_citation=MessageTextFileCitationDetails(
                             file_id=file_id,
-                        )
-                    )
+                        ),
+                    ),
                 )
             finally:
                 await client.vector_stores.delete(file_search_tool.resources.file_search.vector_store_ids[0])
 
     def _get_open_api_tool(self):
         """Helper method to get the openAPI tool."""
-        weather_asset_file_path = os.path.join(
-            os.path.dirname(__file__), "assets", "weather_openapi.json")
+        weather_asset_file_path = os.path.join(os.path.dirname(__file__), "assets", "weather_openapi.json")
         auth = OpenApiAnonymousAuthDetails()
         with open(weather_asset_file_path, "r") as f:
             openapi_weather = jsonref.load(f)
         return OpenApiTool(
-            name="get_weather", spec=openapi_weather, description="Retrieve weather information for a location", auth=auth
+            name="get_weather",
+            spec=openapi_weather,
+            description="Retrieve weather information for a location",
+            auth=auth,
         )
 
     @agentClientPreparer()
@@ -3219,7 +3220,7 @@ class TestAgentClientAsync(TestAgentClientBase):
         """Test Bing grounding tool call in non-streaming Scenario."""
         async with self.create_client(by_endpoint=True, **kwargs) as client:
             model_name = "gpt-4o"
-            openapi_tool = BingGroundingTool(connection_id=kwargs.get('azure_ai_agents_tests_bing_connection_id'))
+            openapi_tool = BingGroundingTool(connection_id=kwargs.get("azure_ai_agents_tests_bing_connection_id"))
 
             await self._do_test_tool(
                 client=client,
@@ -3231,7 +3232,7 @@ class TestAgentClientAsync(TestAgentClientBase):
                 uri_annotation=MessageTextUrlCitationDetails(
                     url="*",
                     title="*",
-                ) 
+                ),
             )
 
     @agentClientPreparer()
@@ -3240,7 +3241,7 @@ class TestAgentClientAsync(TestAgentClientBase):
         """Test Bing grounding tool call in streaming Scenario."""
         async with self.create_client(by_endpoint=True, **kwargs) as client:
             model_name = "gpt-4o"
-            openapi_tool = BingGroundingTool(connection_id=kwargs.get('azure_ai_agents_tests_bing_connection_id'))
+            openapi_tool = BingGroundingTool(connection_id=kwargs.get("azure_ai_agents_tests_bing_connection_id"))
 
             await self._do_test_tool_streaming(
                 client=client,
@@ -3252,7 +3253,7 @@ class TestAgentClientAsync(TestAgentClientBase):
                 uri_annotation=MessageTextUrlCitationDetails(
                     url="*",
                     title="*",
-                ) 
+                ),
             )
 
     async def _do_test_tool(
@@ -3345,7 +3346,7 @@ class TestAgentClientAsync(TestAgentClientBase):
                     if has_annotation:
                         break
                 assert has_annotation, f"The annotation [{uri_annotation.title}]({uri_annotation.url}) was not found."
-        
+
             # Search for the file annotation.
             if file_annotation:
                 has_annotation = False
@@ -3427,16 +3428,20 @@ class TestAgentClientAsync(TestAgentClientBase):
 
                     if isinstance(event_data, MessageDeltaChunk):
                         received_message = True
-                    
+
                     elif isinstance(event_data, ThreadMessage):
                         if event_data.role == MessageRole.AGENT:
                             # Search for the specific URL and title in the message annotation.
                             if not has_uri_annotation:
-                                has_uri_annotation = has_uri_annotation or self._has_url_annotation(event_data, uri_annotation)
-                        
+                                has_uri_annotation = has_uri_annotation or self._has_url_annotation(
+                                    event_data, uri_annotation
+                                )
+
                             # Search for the file annotation.
                             if not has_file_annotation:
-                                has_file_annotation = has_file_annotation or self._has_file_annotation(event_data, file_annotation) 
+                                has_file_annotation = has_file_annotation or self._has_file_annotation(
+                                    event_data, file_annotation
+                                )
 
                     elif isinstance(event_data, RunStepDeltaChunk):
                         if expected_delta_class is not None:
@@ -3467,8 +3472,10 @@ class TestAgentClientAsync(TestAgentClientBase):
                 assert got_expected_delta, f"The delta tool call of type {expected_delta_class} was not found."
                 assert is_completed, "The stream was not completed."
                 assert is_run_step_created, "No run steps were created."
-                
-                assert has_uri_annotation, f"The annotation [{uri_annotation.title}]({uri_annotation.url}) was not found."
+
+                assert (
+                    has_uri_annotation
+                ), f"The annotation [{uri_annotation.title}]({uri_annotation.url}) was not found."
                 assert has_file_annotation, f"The annotation {file_annotation} was not found."
             messages = [message async for message in client.messages.list(thread_id=thread.id)]
             assert len(messages) > 1